% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_map.R
\name{cp_map}
\alias{cp_map}
\title{Run a map with the function, but saves after a given number of execution.}
\usage{
cp_map(
  .x,
  .f,
  ...,
  name = NULL,
  import = NULL,
  workers = NULL,
  fill_method = "default",
  .board = NULL,
  n_checkpoint = NULL,
  notification = NULL
)
}
\description{
Run a map with the function, but saves after a given number of execution.
}
