% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_map_lgl.R
\name{cp_map_lgl}
\alias{cp_map_lgl}
\title{Run a map with the function, but saves after a given number of execution.}
\usage{
cp_map_lgl(
  .x,
  .f,
  ...,
  name = NULL,
  wait = 1,
  workers = NULL,
  fill = TRUE,
  n_checkpoint = 100
)
}
\description{
Run a map with the function, but saves after a given number of execution.
}
