% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currr.R
\name{currify_verb}
\alias{currify_verb}
\title{Create a curryable verb.}
\usage{
currify_verb(verb)
}
\arguments{
\item{verb}{A function that takes .data and args (...)}
}
\description{
currify_verb(foo) creates a function of dots (...).
Passing arguments to this function in turn creates a new function foo(.data, ...)
}
\examples{
filtering = currify_verb(dplyr::filter)
}
