% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currr.R
\name{\%.\%}
\alias{\%.\%}
\title{Function composition operator.}
\usage{
g \%.\% f
}
\arguments{
\item{g}{A function}

\item{f}{A function}
}
\description{
\code{g \%.\% f} creates a function equivalent to g(f(x)) for some input \code{x}.
Meant to emulate infix notation for function composition, $f . g$.
}
\examples{
(length \%.\% unique)(c(1, 1, 2, 2, 3))
}
