% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currr.R
\name{\%;\%}
\alias{\%;\%}
\title{Postfix function composition operator.}
\usage{
f \%;\% g
}
\arguments{
\item{f}{A function}

\item{g}{A function}
}
\description{
\code{f \%;\% g} creates a function equivalent to g(f(x)) (or \code{x |> f() |> g()}) for some input \code{x}.
Unlike conventional function composition, which evaluates right to left, postfix notation operates left to right.
This is similar to the direction of the pipe operator \verb{|>}, except the postfix operator creates a function (without evaluating it).
}
\examples{
(unique \%;\% length)(c(1, 1, 2, 2, 3))
}
