% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_currr_cache.R
\name{remove_currr_cache}
\alias{remove_currr_cache}
\title{Remove currr's intermediate data from the folder.}
\usage{
remove_currr_cache(list = NULL, .board)
}
\description{
Remove currr's intermediate data from the folder.
}
