% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_map.R
\name{saving_map}
\alias{saving_map}
\title{Run a map with the function, but saves after a given number of execution.}
\usage{
saving_map(.ids, .f, name, n_checkpoint = 100, ...)
}
\description{
Run a map with the function, but saves after a given number of execution.
}
