% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_map_nodot.R
\name{saving_map_nodot}
\alias{saving_map_nodot}
\title{Run a map with the function, but saves after a given number of execution. This is an internal function, you are not supposed to use it manually, but can call for background job only if exported. This function differs from saving_map, since it does not have a ... input. This is neccessary because job::job fails if ... is not provided for the cp_map call.}
\usage{
saving_map_nodot(.ids, .f, name, n_checkpoint = 100, currr_folder)
}
\arguments{
\item{.ids}{Placement of .x to work with.}

\item{.f}{Called function.}

\item{name}{Name for saving.}

\item{n_checkpoint}{Number of checkpoints.}

\item{currr_folder}{Folder where cache files are stored.}
}
\value{
No return value, called for side effects
}
\description{
Run a map with the function, but saves after a given number of execution. This is an internal function, you are not supposed to use it manually, but can call for background job only if exported. This function differs from saving_map, since it does not have a ... input. This is neccessary because job::job fails if ... is not provided for the cp_map call.
}
