% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_status.R
\name{update_status}
\alias{update_status}
\title{Update the status of the progress bar.}
\usage{
update_status(name, done = 50, n = 230, chr_length = 80, eta = "120 sec")
}
\arguments{
\item{done}{A number.}

\item{n}{A number.}

\item{chr_length}{A number.}

\item{eta}{A string}

\item{cache}{A number.}
}
\description{
Update the status of the progress bar.
}
\examples{
library(currr)
update_status(done = 50, cache = 40, n = 100, chr_length = 80, eta = "120 sec")
Sys.sleep(.1)
update_status(done = 55, cache = 40, n = 100, chr_length = 80, eta = "120 sec")
Sys.sleep(.1)
update_status(done = 60, cache = 40, n = 100, chr_length = 80, eta = "120 sec")

}
