% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_currr_cache.R
\name{remove_currr_cache}
\alias{remove_currr_cache}
\title{Remove currr's intermediate data from the folder.}
\usage{
remove_currr_cache(list = NULL)
}
\arguments{
\item{list}{A character vector specifying the name of the caches you want to remove (files in .currr.data folder). If empy (default), all caches will be removed.}
}
\value{
No return value, called for side effects
}
\description{
Remove currr's intermediate data from the folder.
}
